refactor: rename component names
