+++ app/components/summaryOverview/SummaryOverviewBox.tsx
@@ -1,49 +1,61 @@
-import SummaryOverviewBoxArray from "./SummaryOverviewBoxArray";
-import SummaryOverviewBoxWrapped from "./SummaryOverviewBoxWrapped";
+import EditButton from "@digitalservicebund/icons/CreateOutlined";
+import { Context } from "~/domains/contexts";
+import Heading from "../Heading";
+import SummaryOverviewBoxItem from "./SummaryOverviewBoxItem";
+import Button from "../Button";
 import { useFlowFormular } from "../form/flowFormularContext";
 
-export type SummaryOverviewBoxProps = {
+type Props = {
   readonly title?: string;
   readonly stepId: string;
-  readonly id: number;
-  readonly fields: string;
+  readonly boxId: number;
+  readonly userData: Context;
+  readonly boxPageFields: string[];
 };
 
 const SummaryOverviewBox = ({
-  title,
-  id,
+  boxId,
   stepId,
-  fields,
-}: SummaryOverviewBoxProps) => {
-  const { userData, validFlowPages } = useFlowFormular();
-
-  if (!validFlowPages[stepId]) {
-    return null;
-  }
-
-  const boxPageFields = fields.split("\n");
-
-  const { isArrayPage } = validFlowPages[stepId];
-
-  if (isArrayPage) {
-    return (
-      <SummaryOverviewBoxArray
-        boxId={id}
-        title={title}
-        arrayBoxPageFields={boxPageFields}
-        stepId={stepId}
-      />
-    );
-  }
+  userData,
+  boxPageFields,
+  title,
+}: Props) => {
+  const { translations, flowId } = useFlowFormular();
 
   return (
-    <SummaryOverviewBoxWrapped
-      boxId={id}
-      stepId={stepId}
-      userData={userData}
-      boxPageFields={boxPageFields}
-      title={title}
-    />
+    <div className="mt-8">
+      <div className="bg-white pt-32 pb-44 px-32">
+        {title && (
+          <Heading
+            text={title}
+            className="mb-16"
+            tagName="p"
+            look="ds-heading-03-bold"
+          />
+        )}
+
+        <dl>
+          {boxPageFields.map((field) => (
+            <SummaryOverviewBoxItem
+              key={`${boxId}-${field}`}
+              fieldName={field}
+              translations={translations}
+              userData={userData}
+            />
+          ))}
+        </dl>
+
+        <Button
+          iconLeft={<EditButton />}
+          href={`${flowId}${stepId}`}
+          look="tertiary"
+          size="large"
+          className="w-fit mt-16"
+        >
+          Bearbeiten
+        </Button>
+      </div>
+    </div>
   );
 };
